made API.md more clear and added some comments to string_parse.h

+++ example/twi_serial_bridge/string_parse.h
@@ -20,8 +20,9 @@ char hexc2int(char c) {
 // but not:
 // 123 4567
 // (the 123 would be interpreted as 12 3; the 4567 would be interpreted as 45 67)
-// note that the first byte consumed is the first entry of the array, etc.
-uint8_t parse_hex_array(char *&p, uint8_t *buffer, uint8_t max_count) {
+// Note that the first byte consumed is the first entry of the array, etc.
+// Advances p to the first non-space, non-hexadecimal character.
+uint8_t parse_hex_array(char *&p, uint8_t *buffer, size_t max_count) {
   if (max_count == 0)
     return 0;
   
@@ -60,6 +61,7 @@ uint8_t parse_hex_array(char *&p, uint8_t *buffer, uint8_t max_count) {
 
 // consumes all hexadecimal characters and the first non-hexadecimal character, returning the
 // 16-bit value defined by the last >= 4 characters read
+// Advances p to the first non-hexadecimal character, skipping the next character if it is a space.
 uint16_t parse_hex16(char *&p) {
   uint16_t v = 0;
 
@@ -83,6 +85,7 @@ uint16_t parse_hex16(char *&p) {
 
 // consumes all hexadecimal characters and the first non-hexadecimal character, returning the
 // 8-bit value defined by the last >= 2 characters read
+// Advances p to the first non-hexadecimal character, skipping the next character if it is a space.
 uint8_t parse_hex8(char *&p) {
   uint8_t v = 0;
 
@@ -104,6 +107,8 @@ uint8_t parse_hex8(char *&p) {
   return v;    
 }
 
+// If (*p == c), returns true and advances p one, another if the next character is a space.
+// Otherwise, returns false and does not alter p.
 bool consume_char_if(char *&p, char c) {
   if (*p != c)
     return false;
